final/requirements.txt
albucore==0.0.17
albumentations==1.4.16
annotated-types==0.7.0
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1720457958366/work/certifi
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
distlib==0.3.8
docker-pycreds==0.4.0
eval_type_backport==0.2.0
exceptiongroup==1.2.2
filelock==3.14.0
flatbuffers==24.3.25
fonttools==4.52.4
gitdb==4.0.11
GitPython==3.1.43
humanfriendly==10.0
idna @ file:///croot/idna_1714398848350/work
imageio==2.34.1
iniconfig==2.0.0
joblib @ file:///croot/joblib_1713976767370/work
kiwisolver==1.4.5
lazy_loader==0.4
matplotlib==3.9.0
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mpmath==1.3.0
netron==7.8.4
networkx==3.3
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp310-cp310-linux_x86_64.whl#sha256=d8cd837ed43e87f77e6efaa08e8de927ca030a1c9c5d04624432d6fb9a74a5ee
onnx==1.16.2
onnx-pytorch==0.1.5
onnxruntime==1.19.0
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.84
packaging==24.0
pillow @ file:///croot/pillow_1714398848491/work
platformdirs==4.2.2
pluggy==1.5.0
protobuf==4.25.3
psutil==5.9.8
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.1.2
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
pytest==8.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests @ file:///croot/requests_1716902831423/work
scikit-image==0.23.2
scikit-learn @ file:///croot/scikit-learn_1714164736727/work
scipy @ file:///croot/scipy_1701295040508/work/dist/scipy-1.11.4-cp310-cp310-linux_x86_64.whl#sha256=a5e5a33a143f627fd4bde622162664c9ea97ff3af56f51c0f4c76ec9d1a340a3
sentry-sdk==2.3.1
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.13.2
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
tifffile==2024.5.22
tomli==2.0.1
torch==1.12.0
torchaudio==0.12.0
torchvision==0.13.0
tqdm==4.66.4
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
urllib3 @ file:///croot/urllib3_1715635851070/work
virtualenv==20.26.2
wandb==0.17.0
